Fixed Packages, moving to a more Go-like project structure
