--- Day1/Day1Main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	T1 "github.com/xathon/AdventOfCodeGo/Day1/task1"
-	T2 "github.com/xathon/AdventOfCodeGo/Day1/task2"
-)
-
-func main() {
-	T1.Calc();
-	T2.Calc();
-}
